Created Repositories for the new tables

+++ back-end/src/main/java/com/cinema/CineConnect/model/DTO/ProductRecord.java
@@ -1,5 +1,6 @@
 package com.cinema.CineConnect.model.DTO;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -7,7 +8,7 @@ public record ProductRecord(
         UUID productId,
         String name,
         String type,
-        Float price,
+        BigDecimal price,
         UUID sessionId,
         List<ProductRecord> addOns
 ) {
